backend/requirements.txt
fastapi
uvicorn
motor
celery
redis
python-multipart
python-jose
passlib
argon2-cffi
httpx
beautifulsoup4
pytest
pydantic[email]
pydantic-settings
email-validator
requests
aiohttp
selenium
webdriver-manager
loguru
pandas
numpy
transformers
langchain-core
groq
tavily-python
python-dotenv
bcrypt
